pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptrek"
version = "0.0.7"
description = "Universal AI Editor prompt storage solution"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Addie Rudy", email = "developer@flamingquaks.dev"}
]
keywords = ["ai", "prompts", "copilot", "cursor", "continue", "cli"]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Utilities",
]

dependencies = [
    "click>=8.0.0",
    "PyYAML>=6.0",
    "Jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "isort>=6.0.1",
    "black>=25.9.0",
    "flake8>=7.3.0",
]

[project.scripts]
promptrek = "promptrek.cli.main:cli"

[project.urls]
Homepage = "https://github.com/flamingquaks/promptrek"
Repository = "https://github.com/flamingquaks/promptrek"
Documentation = "https://github.com/flamingquaks/promptrek/tree/main/docs"
Issues = "https://github.com/flamingquaks/promptrek/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
promptrek = ["../.pre-commit-hooks.yaml"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
    "*.egg-info"
]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
exclude = ["tests/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["src"]
addopts = [
    "--cov=promptrek",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]

[tool.coverage.run]
source = ["src/promptrek"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "isort>=5.0.0",
    "types-pyyaml>=6.0.12.20241230",
    "pre-commit>=3.0.0",
    "yamllint>=1.35.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "isort>=5.0.0",
    "types-pyyaml>=6.0.12.20241230",
    "pre-commit>=3.0.0",
    "yamllint>=1.35.0",
    "build>=1.3.0",
]
